moved can_edit to core
