+++ client/templates/files/cellmlmodel_detail.html
@@ -1,5 +1,4 @@
 {% extends "base.html" %}
-{% load files %}
 
 {% block title %}CellML model - {{ object }}{% endblock title %}
