+++ client/templates/simulations/simulation_header.html
@@ -1,5 +1,5 @@
 {% load static %}
-{% load simulations %}
+{% load ap_nimbus %}
 
 {% if object.id %}
 <a href="{% url 'simulations:simulation_edit' object.id %}"><h1 id="entityname">Simulation: {{ object.title }}.</h1></a>
